JSON: Move json decoding function to alarm node
