+++ ros_alarms/__init__.py
@@ -1,4 +1,5 @@
 from alarms import AlarmListener, AlarmBroadcaster, HeartbeatMonitor
+from alarms import parse_json_str
 
 try:
     from tx_alarms import TxAlarmListener, TxAlarmBroadcaster, TxHeartbeatMonitor
